Video: Now encapsulating some more stuff (resolution, ratio, ...) Video: Tried fixing and cleaning up resizing. This still does not really work on OSX.
ModelManager, Model: Now changing file ending in manager rather than constructor of Model to get the same entry for A.mdx, A.m2 and A.mdl.

MapChunk, MapView, MapTile, World: Changed selection.
MapView: doSelection now takes a bool indicating if only terrain should be selected rather than an integer with nonsense name.
MapView: doSelection is now only calling gWorld->drawSelection.
MapView: Instead of checking AutoSelection as well as mode and setting AutoSelection at thousands of places, in displayViewMode_3D is now always doing selection, if no static object is selected and autoselectingmode is true.
MapView: Rather than setting a variable per mouse key to false and then checking for it being false (which it could be as well if it did not just get released), the confusing code is now also in the switch for which button is released.
Environment: Removed AutoSelection. This is not needed and is confusing with having AutoSelectionMode (mode being bool..) as well.
World: eSelectionMode removed.
World: getSelection is now private.
World: Selecting now first pushes MapObjName, DoodadName or MapTileName on the stack, then draws these. This is made to be able to see instantly, what type of item got selected. Also, this enables us to do the sub selection of the triangle at the same time as the chunk, saving one draw.
World: The returned name stack is now traversed in a readable way.
TODO: Redo nameEntry, now being partly obsolete or bad.
MapChunk: drawSelect2() removed. Now pushing both, the Chunk name and the triangle name on the stack.

Settings: FarZ is NOT limited to 1024, 2048, 3072 (not being a power of two anyway..) and 4096, but can actually be everything.
World: Removed never used uselowlod.