(Possible) Fix for strength save bug
https://youtu.be/cCoSve4mS2w
As of the video above, sometimes when using a strength below max, saving will cause weird dots around the opacity.
It seems that it is generally caused by the paintTexture function being able to create situations where the layers don't sum up to 1, and a divide by zero sometimes causing a -nan that also messes with the math. 
I've changed to check cases where painting would go beyond possible values, as well as ensuring that ground layer is always 1 - sumofotherlayers, as that's how the ADT assumes it works.

This should hopefully fix most of those cases.